requirements.txt
pillow
streamlit
openai
